Javascript example of the API
This is a Javascript example of the API. To make it work, you must first register a .pfx certificate in your navigator.
On Chrome, go in Settings - Scroll down to security - select Manage Certificates - import the certificate. You will be prompted for a password. The password is the same has for the ZIP file.